Move processMetadataAndSaveAsPhoto to photoservice
